Make infra script more general

+++ bin/infra.sh
@@ -2,8 +2,6 @@
 
 cd /
 
-USR="usr"
-
 # Might run at first boot, services might be already running
 
 echo "PasswordAuthentication no" >> /etc/ssh/sshd_config
@@ -15,36 +13,41 @@ echo "Port ${SSHDPORT}" >> /etc/ssh/sshd_config
 
 systemctl restart sshd
 
-adduser --disabled-password --gecos "" $USR
-usermod -aG sudo $USR
-usermod -aG adm $USR
-
-rm -rf /home/$USR/.*
-mkdir -p /home/$USR/.ssh/
-
-# Take password from root
-sed -i "/^usr:/d" /etc/shadow
-head -1 /etc/shadow | sed -e "s/^root/usr/" >> /etc/shadow
-
-# Take key from root
-mv /root/.ssh/authorized_keys /home/$USR/.ssh/
-chmod 400 /home/$USR/.ssh/authorized_keys
-chown -R $USR:$USR /home/$USR/.ssh/
-rm -rf /root/.ssh
-
-# Disable root login
-usermod -p '*' root
-
-echo "RESUME=none" > /etc/initramfs-tools/conf.d/noresume.conf
+if [ -d /etc/initramfs-tools ]; then
+  echo "RESUME=none" > /etc/initramfs-tools/conf.d/noresume.conf
+fi
 
 apt-get update
-apt-get install -y -qq --no-install-recommends git
 
-runuser -u $USR -- git clone https://github.com/gombos/dotfiles.git /home/$USR/.dotfiles
-runuser -u $USR -- /home/$USR/.dotfiles/bin/infra-provision-user.sh
+if ! [ -z "$USR" ]; then
+  adduser --disabled-password --gecos "" $USR
+  usermod -aG sudo $USR
+  usermod -aG adm $USR
 
-if [ -d /home/$USR/.dotfiles/bin ]; then
-  PATH=/home/$USR/.dotfiles/bin:$PATH
+  rm -rf /home/$USR/.*
+  mkdir -p /home/$USR/.ssh/
+
+  # Take password from root
+  sed -i "/^usr:/d" /etc/shadow
+  head -1 /etc/shadow | sed -e "s/^root/usr/" >> /etc/shadow
+
+  # Take key from root
+  mv /root/.ssh/authorized_keys /home/$USR/.ssh/
+  chmod 400 /home/$USR/.ssh/authorized_keys
+  chown -R $USR:$USR /home/$USR/.ssh/
+  rm -rf /root/.ssh
+
+  # Disable root login
+  usermod -p '*' root
+
+  apt-get install -y -qq --no-install-recommends git
+
+  runuser -u $USR -- git clone https://github.com/gombos/dotfiles.git /home/$USR/.dotfiles
+  runuser -u $USR -- /home/$USR/.dotfiles/bin/infra-provision-user.sh
+
+  if [ -d /home/$USR/.dotfiles/bin ]; then
+    PATH=/home/$USR/.dotfiles/bin:$PATH
+  fi
 fi
 
 # maybe call usrlocal script
@@ -56,7 +59,7 @@ apt-get install -y -qq --no-install-recommends unzip micro
 # Takes time, do it last
 apt-mark hold linux-image-amd64
 
-apt-get purge -y -q rsyslog telnet traceroute os-prober tasksel javascript-common vim-* whiptail publicsuffix nano dmidecode hdparm iso-codes mtr-tiny pciutils reportbug whois gcc-9-base libnewt0.52 libgpm2 libldap-common liblockfile-bin libsasl2-modules curl dnsutils apt-listchanges libslang2 liblognorm5 debconf-i18n
+apt-get purge -y -q rsyslog telnet traceroute os-prober tasksel javascript-common vim-* whiptail publicsuffix nano mtr-tiny reportbug whois libldap-common liblockfile-bin libsasl2-modules dnsutils apt-listchanges liblognorm5 debconf-i18n 2>/dev/null >/dev/null
 
 # Ubuntu things
 # apt-mark hold linux-image-generic
@@ -82,7 +85,7 @@ apt-get -y -qq upgrade
 
 #infra-clean-linux.sh
 
-if [ -z "$LABEL" ]; then
+if ! [ -z "$LABEL" ]; then
   hostnamectl set-hostname ${LABEL}
 fi
